stats calculation only in useMathProblems
